Refactor JWT token code
